Working on dvb subtitles
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@2354 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/TransportStream/DvbSubtitle.cs
@@ -7,8 +7,8 @@ namespace Nikse.SubtitleEdit.Logic.TransportStream
 {
     public class DvbSubtitle
     {
-        public int StartMilliseconds { get; set; }
-        public int EndMilliseconds { get; set; }
+        public ulong StartMilliseconds { get; set; }
+        public ulong EndMilliseconds { get; set; }
         public DvbSubPes Pes { get; set; }      
 
         /// <summary>
